Rework remember me mechanism (fix time calculation) #2163

+++ modules/core/src/com/haulmont/cuba/security/auth/providers/RememberMeAuthenticationProvider.java
@@ -140,7 +140,7 @@ public class RememberMeAuthenticationProvider extends AbstractAuthenticationProv
 
         long tokenCreated = rememberMeToken.getCreateTs().getTime();
         long now = timeSource.currentTimeMillis();
-        int expirationTimeout = globalConfig.getRememberMeExpirationTimeoutSec();
+        long expirationTimeout = globalConfig.getRememberMeExpirationTimeoutSec();
 
         return tokenCreated + expirationTimeout * 1000 < now;
     }
